requirements.txt
torch
torchvision
transformers
albumentations
opencv-python
pillow
numpy
matplotlib
scikit-learn
streamlit
tqdm
easyocr
pandas
python-dateutil
pymupdf